search: Handle package::requires and component::requires correctly
This correctly implements the algorithm to use package::requires as a
set of required components with restrictions, but the final graph only
has the requirements that individual components have.

There isn't an elegant way to avoid doing two passes, because of the
diamond problem. A requires B and C, B requires D:FOO, C requires D:BAR,
we cannot trim D after B because then it will no longer be valid for D,
we must wait until the full graph is built.