LICENSE: add a License file with the MIT license
This helps github to identify the license of the project and display it
in the web interface.

I copied this from the OSI website (https://opensource.org/license/mit),
but without the `Copyright <year> <holder>`, since the project does not
have an entity holding the license, and individual files may have
different copyright holders.

Fixes: #25