Prefix all in-source references to libcps
**Including**

* Spell `#include` statements like `cps/*.hpp`
* Spell preprocessor definitions like `CPS_*`
* Made enumerated values `lowercase` instead of
  `UPPERCASE` to avoid collisions with preprocessor
  definitions from other projects.

**Incidental Changes**

To make this commit less awkward, some incidental design changes
were scoped in, including adding a `src/cps` subdirectory along with
`meson.build` files to describe that subdirectory in a more
encapsulated way for the rest of the build configuration.

`HAVE_UNREACHABLE` was renamed to `CPS_USE_BUILTIN_UNREACHABLE`.

`main` was reimplemented to be a pass-through call to a new
`cps_config::run` function. This added a namespace to the
`cps-config` program itself that is distinct from the namespace
for the `cps` library.

Incidental changes to include sorting, quoting, and spacing
were added to make automated code formatters behave well
with current include sorting and spelling patterns.

Updated some copyright notices.

**Justification**

This will better prepare the project to publish a `libcps` in the future,
which will be desireable.

Specifically:

* These changes will avoid conflicts with other projects that might ship,
  for instance, a `utils.hpp`
* These changes will avoid collisions with other projects that might
  define a variable named `unreachable`.
* These changes will avoid collisins with other projects that might
  define preprocessor macros like `SIMPLE`.