CI: fix ASan on Ubuntu
The GitHub runner Ubuntu image has a Kernel configured with stronger
than default ASLR settings. This breaks the older versions of LLVM they
ship, including libasan. To fix this we can set the ASLR settings back to
default. This is supposed to be fixed in the next GitHub image, at which
point the workaround ca be removed.

See: https://github.com/actions/runner-images/issues/9491