Use Meson waps on Rocky Linux
This change intentionally takes on a little technical debt in
order to allow productive development of this project.

As noted in the comment in this commit, the main issue is an API
incompatability across the versions of cxxopts on the two OSs.
Instead of this approach, it would be possible to instead add
preprocessor logic to support both APIs in the same source file.
That is a valid approach, but this commit instead values keeping
the C++ code straightforward. Preprocessor tricks can be
brittle, tricky to test, confusing to humans, and confusing
to tools.