Switch from cxxopts to CLI11
Change from using cxxopts for command-line parsing to using
CLI11.

Use lambda passing syntax for non-trivial flag handling so as
to put the declaration of the flag next to the implementation for
that flag.

Use the CLI11 footer feature to move documentation for examples
and support to the bottom of the help message.

Support more than one value for packages and components.

Closes #76.