Rerun actions on workflow changes
A previous commit [1] was made to limit when the tests are run. However,
the list in [1] does not include the workflow file itself. This prevents
tests from being executed if a commit or PR only modifies test.yml
workflow file. That is, without this change, this very PR would not
trigger any testing.

[1] https://github.com/cps-org/cps-config/commit/b901168e1617fd5c15f30a65b62126fda594339b