Merge package installation into a single layer
Currently, CI is failing weirdly because of ubuntu repos returning 404.
I suspect that is because the package index layer, `apt-get update`, is
a cache-hit but the next layer that installs packages are not. This
causes `apt` to work with outdated indices.

Further, it's just good container hygiene to keep those in a single
layer to make sure package index doesn't end up in a layer of the image.