Add initial scanner and parser for pkg-config file
This commit implements a simple parser for pkg-config files. This
implementation supports variable assignments and interpolation. It also
supports comments. The grammar documented here [1] is used as a
reference.

This implementation has the following intentional incorrectness and
omissions:
1. All consecutive whitespaces in fragment lists and literal values are
replaced by a single string. I'm not sure whether this will affect
anything at all, so I made this simplification to allow for more
succinct grammar.
2. Version requirements in dependency lists are intentionally omitted
and left as future work.
3. This implementation only contains a class to load PC files, but that
is not used anywhere. To keep this PR minimal, the logic to fall back
to pc files when cps files are not found is also left as future work.
4. Also as future work, the ability to parse pc files should probably be
optional.

[1] https://manpages.ubuntu.com/manpages/jammy/man5/pc.5.html