loader: fix loading `package::requires` children
Due to a bug in the loader we would try to search for the sub elements
from `requires` in `package`, which would not be found. This was covered
up by a test that had the cps file incorrectly specified to include the
`.cps` extension, so we looked up `needs-version.cps.cps`, which doesn't
exist, and then checked for the error message associated with that.