Fix interaction with prefix calculation
In the current form, filename is default constructed to empty string and
causes calculate prefix to fail. Setting cps_path to nullopt and passing
filename as expected solves this issue.

In the future, PREFIX variable in pc files should be given special
treatment and be translated to @prefix@ in cps.