search: handle arbitrarily deep self requires references
This requires that all requirements of the current node be processed
before starting on any of its children, otherwise we may not select all
of the correct components from those children.

I still don't think the algorithm is fundamentally correct, since we
could still end up visiting the same child more than once, and could
fail to get all of the requirements we need for each child before
visiting it. However, this is an improvement over the status quo, and
does fix the linked issue.

Fixes: #101