tests: import pkgconf tests
There is still some work to do sorting out with upstream some tests that
are defined but not run, there are also a few tests not ported (due to
complexity of the port) but in general this is close enough to merge.
There's about 100 passes and 60 failures currently, the failures have
been marked as expected failures, and the test runner understands these
are expected failures (annotated in the toml) and reports them as passes
as long as they fail.

This is currently not plugged into the cmake.